pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "listingjet"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "PyJWT>=2.8",
    "cryptography>=42",
    "passlib[bcrypt]>=1.7",
    "temporalio>=1.7",
    "redis[hiredis]>=5.0",
    "boto3>=1.34",
    "httpx>=0.27",
    "anthropic>=0.28",
    "stripe>=8.0",
    "Pillow>=10.0",
    "sentry-sdk>=1.40",
    "imagehash>=4.3",
    "clamd>=1.0.2",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "factory-boy>=3.3",
    "coverage>=7.5",
    "moto[s3]>=5.0",
    "fakeredis>=2.23",
    "ruff>=0.8",
    "locust>=2.20",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/launchlens"]

[tool.coverage.run]
source = ["src/launchlens"]
omit = ["*/test*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 60
show_missing = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
